+++ R/osmdata-class.R
@@ -15,7 +15,6 @@
 #' @param osm_multipolygons OSM relations as \pkg{sf} Simple Features
 #'                          Collection of multipolygons or \pkg{sp}
 #'                          SpatialPolygonsDataFrame 
-#' @param ... other options ignored
 #'
 #' @note Class constructor should never be used directly, and is only exported
 #' to provide access to the print method